+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ## Version 0.52.0
 **Release Date**: 21st June 2023
 ### Features
-1. Add Integration Instance
+1. Add Integration Instance Resource
+2. Add Instance Configuration Resource for OCI
 ### Bug Fixes
 1. Fix incorrect Markdown generation when markdown view displayed
 2. Resolve issue where PCA Loadbalancers were not linking all backend instances.